pyproject.toml
[tool.poetry]
name = "analysis-banking-transactions"
version = "0.1.0"
description = ""
authors = ["Anton Ershov <ant_e@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.4"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.0"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119

[tool.black]
line-length = 119

[tool.mypy]
warn_return_any = true
no_implicit_optional = true
disallow_untyped_defs = true
exclude = 'env'